test: 당첨 통계 로직 - Lottos -> LottoRank 역할 변경
